backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.12.1
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
filelock==3.19.1
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2025.10.0
google-ai-generativelanguage==0.6.6
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.7.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.0
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==1.35.0
packaging==26.1
proto-plus==1.27.2
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
rich==15.0.0
s3transfer==0.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
Werkzeug==3.1.8
zipp==3.23.1
